Fix state reset
Some regs were left alone by the reset, which made Xs propagate and
made the output unreliable.
Moved some signal updates that were being done on the posedge
of 'scl' into the clk domain, to make the updates more reliable. This
was probably the reason for the ACK not working properly.